+++ lib/bluetooth/devices/base_device.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:dartx/dartx.dart';
 import 'package:flutter/foundation.dart';
@@ -43,13 +42,21 @@ abstract class BaseDevice {
 
   static BaseDevice? fromScanResult(BleDevice scanResult) {
     // Use the name first as the "System Devices" and Web (android sometimes Windows) don't have manufacturer data
-    final device = switch (scanResult.name) {
-      //'Zwift Ride' => ZwiftRide(scanResult), special case for Zwift Ride: we must only connect to the left controller
-      // https://www.makinolo.com/blog/2024/07/26/zwift-ride-protocol/
-      'Zwift Play' => ZwiftPlay(scanResult),
-      //'Zwift Click' => ZwiftClick(scanResult), special case for Zwift Click v2: we must only connect to the left controller
-      _ => null,
-    };
+    final device =
+        kIsWeb
+            ? switch (scanResult.name) {
+              'Zwift Ride' => ZwiftRide(scanResult),
+              'Zwift Play' => ZwiftPlay(scanResult),
+              'Zwift Click' => ZwiftClickV2(scanResult),
+              _ => null,
+            }
+            : switch (scanResult.name) {
+              //'Zwift Ride' => ZwiftRide(scanResult), special case for Zwift Ride: we must only connect to the left controller
+              // https://www.makinolo.com/blog/2024/07/26/zwift-ride-protocol/
+              'Zwift Play' => ZwiftPlay(scanResult),
+              //'Zwift Click' => ZwiftClick(scanResult), special case for Zwift Click v2: we must only connect to the left controller
+              _ => null,
+            };
 
     if (device != null) {
       return device;
@@ -101,7 +108,7 @@ abstract class BaseDevice {
 
     await UniversalBle.connect(device.deviceId);
 
-    if (!kIsWeb && Platform.isAndroid) {
+    if (!kIsWeb) {
       await UniversalBle.requestMtu(device.deviceId, 517);
     }
 
@@ -151,10 +158,10 @@ abstract class BaseDevice {
     await UniversalBle.subscribeNotifications(device.deviceId, customService.uuid, asyncCharacteristic.uuid);
     await UniversalBle.subscribeIndications(device.deviceId, customService.uuid, syncTxCharacteristic.uuid);
 
-    await _setupHandshake();
+    await setupHandshake();
   }
 
-  Future<void> _setupHandshake() async {
+  Future<void> setupHandshake() async {
     if (supportsEncryption) {
       await UniversalBle.write(
         device.deviceId,
@@ -181,10 +188,9 @@ abstract class BaseDevice {
   Future<void> processCharacteristic(String characteristic, Uint8List bytes) async {
     if (kDebugMode) {
       print(
-        '${DateTime.now().toString().split(" ").last} Received $characteristic: ${bytes.map((e) => e.toRadixString(16).padLeft(2, '0')).join(' ')} => ${String.fromCharCodes(bytes)} ',
+        "${DateTime.now().toString().split(" ").last} Received data on $characteristic: ${bytes.map((e) => e.toRadixString(16).padLeft(2, '0')).join(' ')}",
       );
     }
-
     if (bytes.isEmpty) {
       return;
     }
@@ -192,27 +198,8 @@ abstract class BaseDevice {
     try {
       if (bytes.startsWith(startCommand)) {
         _processDevicePublicKeyResponse(bytes);
-      } else if (bytes.startsWith(Constants.RIDE_ON)) {
-        //print("Empty RideOn response - unencrypted mode");
-        if (this is ZwiftClickV2) {
-          // TODO figure out if this is the key to make V2 work
-          await UniversalBle.write(
-            device.deviceId,
-            customServiceId,
-            syncRxCharacteristic!.uuid,
-            Uint8List.fromList([0x00, 0x08, 0x00]),
-            withoutResponse: true,
-          );
-          await UniversalBle.write(
-            device.deviceId,
-            customServiceId,
-            syncRxCharacteristic!.uuid,
-            Uint8List.fromList([0x41, 0x05, 0x05]),
-            withoutResponse: true,
-          );
-        }
       } else if (!supportsEncryption || (bytes.length > Int32List.bytesPerElement + EncryptionUtils.MAC_LENGTH)) {
-        _processData(bytes);
+        processData(bytes);
       }
     } catch (e, stackTrace) {
       print("Error processing data: $e");
@@ -233,7 +220,7 @@ abstract class BaseDevice {
     zapEncryption.initialise(devicePublicKeyBytes);
   }
 
-  Future<void> _processData(Uint8List bytes) async {
+  Future<void> processData(Uint8List bytes) async {
     int type;
     Uint8List message;
 
@@ -258,59 +245,7 @@ abstract class BaseDevice {
       message = bytes.sublist(1);
     }
 
-    if (bytes.startsWith(Constants.RESPONSE_STOPPED_CLICK_V2)) {
-      print('no more events');
-      //connect();
-    }
-
     switch (type) {
-      case Constants.UNKNOWN_CLICKV2_TYPE:
-        if (!_isInited) {
-          _isInited = true;
-          final commands = [
-            [0x00, 0x08, 0x80, 0x08],
-            [0x00, 0x08, 0x83, 0x06],
-            [
-              0xFF,
-              0x04,
-              0x00,
-              0x0A,
-              0x15,
-              0x40,
-              0xE9,
-              0xD9,
-              0xC9,
-              0x6B,
-              0x74,
-              0x63,
-              0xC2,
-              0x7F,
-              0x1B,
-              0x4E,
-              0x4D,
-              0x9F,
-              0x1C,
-              0xB1,
-              0x20,
-              0x5D,
-              0x88,
-              0x2E,
-              0xD7,
-              0xCE,
-            ],
-          ];
-          for (final command in commands) {
-            await UniversalBle.write(
-              device.deviceId,
-              customServiceId,
-              syncRxCharacteristic!.uuid,
-              Uint8List.fromList(command),
-              withoutResponse: true,
-            );
-            await Future.delayed(const Duration(milliseconds: 50));
-          }
-        }
-        break;
       case Constants.EMPTY_MESSAGE_TYPE:
         //print("Empty Message"); // expected when nothing happening
         break;
@@ -322,47 +257,10 @@ abstract class BaseDevice {
         break;
       case Constants.CLICK_NOTIFICATION_MESSAGE_TYPE:
       case Constants.PLAY_NOTIFICATION_MESSAGE_TYPE:
-      case Constants.RIDE_NOTIFICATION_MESSAGE_TYPE: // untested
+      case Constants.RIDE_NOTIFICATION_MESSAGE_TYPE:
         processClickNotification(message)
             .then((buttonsClicked) async {
-              if (buttonsClicked == null) {
-                // ignore, no changes
-              } else if (buttonsClicked.isEmpty) {
-                actionStreamInternal.add(LogNotification('Buttons released'));
-                _longPressTimer?.cancel();
-
-                // Handle release events for long press keys
-                final buttonsReleased = _previouslyPressedButtons.toList();
-                if (buttonsReleased.isNotEmpty) {
-                  await _performRelease(buttonsReleased);
-                }
-                _previouslyPressedButtons.clear();
-              } else {
-                // Handle release events for buttons that are no longer pressed
-                final buttonsReleased = _previouslyPressedButtons.difference(buttonsClicked.toSet()).toList();
-                if (buttonsReleased.isNotEmpty) {
-                  await _performRelease(buttonsReleased);
-                }
-
-                final isLongPress =
-                    buttonsClicked.singleOrNull != null &&
-                    actionHandler.supportedApp?.keymap.getKeyPair(buttonsClicked.single)?.isLongPress == true;
-
-                if (!isLongPress &&
-                    !(buttonsClicked.singleOrNull == ZwiftButton.onOffLeft ||
-                        buttonsClicked.singleOrNull == ZwiftButton.onOffRight)) {
-                  // we don't want to trigger the long press timer for the on/off buttons, also not when it's a long press key
-                  _longPressTimer?.cancel();
-                  _longPressTimer = Timer.periodic(const Duration(milliseconds: 250), (timer) async {
-                    _performActions(buttonsClicked, true);
-                  });
-                } else if (isLongPress) {
-                  // Update currently pressed buttons
-                  _previouslyPressedButtons = buttonsClicked.toSet();
-                }
-
-                return _performActions(buttonsClicked, false);
-              }
+              return handleButtonsClicked(buttonsClicked);
             })
             .catchError((e) {
               actionStreamInternal.add(LogNotification(e.toString()));
@@ -373,9 +271,51 @@ abstract class BaseDevice {
 
   Future<List<ZwiftButton>?> processClickNotification(Uint8List message);
 
+  Future<void> handleButtonsClicked(List<ZwiftButton>? buttonsClicked) async {
+    if (buttonsClicked == null) {
+      // ignore, no changes
+    } else if (buttonsClicked.isEmpty) {
+      actionStreamInternal.add(LogNotification('Buttons released'));
+      _longPressTimer?.cancel();
+
+      // Handle release events for long press keys
+      final buttonsReleased = _previouslyPressedButtons.toList();
+      if (buttonsReleased.isNotEmpty) {
+        await _performRelease(buttonsReleased);
+      }
+      _previouslyPressedButtons.clear();
+    } else {
+      // Handle release events for buttons that are no longer pressed
+      final buttonsReleased = _previouslyPressedButtons.difference(buttonsClicked.toSet()).toList();
+      if (buttonsReleased.isNotEmpty) {
+        await _performRelease(buttonsReleased);
+      }
+
+      final isLongPress =
+          buttonsClicked.singleOrNull != null &&
+          actionHandler.supportedApp?.keymap.getKeyPair(buttonsClicked.single)?.isLongPress == true;
+
+      if (!isLongPress &&
+          !(buttonsClicked.singleOrNull == ZwiftButton.onOffLeft ||
+              buttonsClicked.singleOrNull == ZwiftButton.onOffRight)) {
+        // we don't want to trigger the long press timer for the on/off buttons, also not when it's a long press key
+        _longPressTimer?.cancel();
+        _longPressTimer = Timer.periodic(const Duration(milliseconds: 250), (timer) async {
+          _performActions(buttonsClicked, true);
+        });
+      } else if (isLongPress) {
+        // Update currently pressed buttons
+        _previouslyPressedButtons = buttonsClicked.toSet();
+      }
+
+      return _performActions(buttonsClicked, false);
+    }
+  }
+
   Future<void> _performActions(List<ZwiftButton> buttonsClicked, bool repeated) async {
     if (!repeated &&
-        buttonsClicked.any(((e) => e.action == InGameAction.shiftDown || e.action == InGameAction.shiftUp))) {
+        buttonsClicked.any(((e) => e.action == InGameAction.shiftDown || e.action == InGameAction.shiftUp)) &&
+        settings.getVibrationEnabled()) {
       await _vibrate();
     }
     for (final action in buttonsClicked) {