+++ lib/bluetooth/devices/zwift_ride.dart
@@ -1,10 +1,15 @@
-import 'dart:typed_data';
-
+import 'package:dartx/dartx.dart';
+import 'package:flutter/foundation.dart';
+import 'package:protobuf/protobuf.dart' as $pb;
 import 'package:swift_control/bluetooth/devices/base_device.dart';
 import 'package:swift_control/bluetooth/messages/ride_notification.dart';
 import 'package:swift_control/utils/keymap/buttons.dart';
+import 'package:universal_ble/universal_ble.dart';
 
+import '../../main.dart';
 import '../ble.dart';
+import '../messages/notification.dart';
+import '../protocol/zp.pb.dart';
 
 class ZwiftRide extends BaseDevice {
   ZwiftRide(super.scanResult)
@@ -40,6 +45,114 @@ class ZwiftRide extends BaseDevice {
   RideNotification? _lastControllerNotification;
 
   @override
+  Future<void> processData(Uint8List bytes) async {
+    Opcode? opcode;
+    Uint8List message;
+
+    if (supportsEncryption) {
+      final counter = bytes.sublist(0, 4); // Int.SIZE_BYTES is 4
+      final payload = bytes.sublist(4);
+
+      if (zapEncryption.encryptionKeyBytes == null) {
+        actionStreamInternal.add(
+          LogNotification(
+            'Encryption not initialized, yet. You may need to update the firmware of your device with the Zwift Companion app.',
+          ),
+        );
+        return;
+      }
+
+      final data = zapEncryption.decrypt(counter, payload);
+      opcode = Opcode.valueOf(data[0]);
+      message = data.sublist(1);
+    } else {
+      opcode = Opcode.valueOf(bytes[0]);
+      message = bytes.sublist(1);
+    }
+
+    if (bytes.startsWith(Constants.RESPONSE_STOPPED_CLICK_V2)) {
+      actionStreamInternal.add(
+        LogNotification('Your Zwift Click V2 no longer sends events. Connect it in the Zwift app once per day.'),
+      );
+    }
+
+    switch (opcode) {
+      case Opcode.RIDE_ON:
+        //print("Empty RideOn response - unencrypted mode");
+
+        break;
+      case Opcode.STATUS_RESPONSE:
+        final status = StatusResponse.fromBuffer(message);
+        if (kDebugMode) {
+          print('StatusResponse: ${status.command} status: ${Status.valueOf(status.status)}');
+        }
+        break;
+      case Opcode.GET_RESPONSE:
+        final response = GetResponse.fromBuffer(message);
+        final dataObjectType = DO.valueOf(response.dataObjectId);
+        if (kDebugMode) {
+          print('GetResponse: ${dataObjectType?.value.toRadixString(16).padLeft(4, '0')} $dataObjectType');
+        }
+
+        switch (dataObjectType) {
+          case DO.PAGE_DEV_INFO:
+            final pageDevInfo = DevInfoPage.fromBuffer(response.dataObjectData);
+            if (kDebugMode) {
+              print('PageDevInfo: $pageDevInfo');
+            }
+            break;
+          case DO.PAGE_DATE_TIME:
+            final pageDateTime = DateTimePage.fromBuffer(response.dataObjectData);
+            if (kDebugMode) {
+              print('PageDateTime: $pageDateTime');
+            }
+            break;
+          case DO.PAGE_CONTROLLER_INPUT_CONFIG:
+            final pageDateTime = ControllerInputConfigPage.fromBuffer(response.dataObjectData);
+            if (kDebugMode) {
+              print('PageDateTime: $pageDateTime');
+            }
+            break;
+          default:
+            break;
+        }
+        break;
+      case Opcode.VENDOR_MESSAGE:
+        break;
+      case Opcode.LOG_DATA:
+        final logMessage = LogDataNotification.fromBuffer(message);
+        if (kDebugMode) {
+          actionStreamInternal.add(LogNotification(logMessage.toString()));
+        }
+        break;
+      case Opcode.BATTERY_NOTIF:
+        final notification = BatteryNotification.fromBuffer(message);
+        if (batteryLevel != notification.newPercLevel) {
+          batteryLevel = notification.newPercLevel;
+          connection.signalChange(this);
+        }
+        break;
+      case Opcode.CONTROLLER_NOTIFICATION:
+        processClickNotification(message)
+            .then((buttonsClicked) async {
+              return handleButtonsClicked(buttonsClicked);
+            })
+            .catchError((e) {
+              actionStreamInternal.add(LogNotification(e.toString()));
+            });
+        break;
+      case null:
+        if (bytes[0] == 0x1A) {
+          final batteryStatus = BatteryStatus.fromBuffer(message);
+          if (kDebugMode) {
+            print('BatteryStatus: $batteryStatus');
+          }
+        }
+        break;
+    }
+  }
+
+  @override
   Future<List<ZwiftButton>?> processClickNotification(Uint8List message) async {
     final RideNotification clickNotification = RideNotification(message);
     if (_lastControllerNotification == null || _lastControllerNotification != clickNotification) {
@@ -52,5 +165,33 @@ class ZwiftRide extends BaseDevice {
     } else {
       return null;
     }
+  }
+
+  Future<void> sendCommand(Opcode opCode, $pb.GeneratedMessage message) async {
+    final buffer = Uint8List.fromList([opCode.value, ...message.writeToBuffer()]);
+    if (kDebugMode) {
+      print("Sending $opCode: ${buffer.map((e) => e.toRadixString(16).padLeft(2, '0')).join(' ')}");
+    }
+    await UniversalBle.write(
+      device.deviceId,
+      customServiceId,
+      syncRxCharacteristic!.uuid,
+      buffer,
+      withoutResponse: true,
+    );
+    await Future.delayed(Duration(milliseconds: 500));
+  }
+
+  Future<void> sendCommandBuffer(Uint8List buffer) async {
+    if (kDebugMode) {
+      print("Sending ${buffer.map((e) => e.toRadixString(16).padLeft(2, '0')).join(' ')}");
+    }
+    await UniversalBle.write(
+      device.deviceId,
+      customServiceId,
+      syncRxCharacteristic!.uuid,
+      buffer,
+      withoutResponse: true,
+    );
   }
 }