+++ lib/pages/requirements.dart
@@ -3,8 +3,10 @@ import 'dart:io';
 import 'package:dartx/dartx.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 import 'package:swift_control/main.dart';
 import 'package:swift_control/utils/requirements/platform.dart';
+import 'package:swift_control/widgets/changelog_dialog.dart';
 import 'package:swift_control/widgets/menu.dart';
 import 'package:swift_control/widgets/title.dart';
 
@@ -30,6 +32,7 @@ class _RequirementsPageState extends State<RequirementsPage> with WidgetsBinding
     // call after first frame
     WidgetsBinding.instance.addPostFrameCallback((_) {
       settings.init().then((_) {
+        _checkAndShowChangelog();
         if (!kIsWeb && Platform.isMacOS) {
           // add more delay due to CBManagerStateUnknown
           Future.delayed(const Duration(seconds: 2), () {
@@ -46,6 +49,23 @@ class _RequirementsPageState extends State<RequirementsPage> with WidgetsBinding
         Navigator.push(context, MaterialPageRoute(builder: (c) => DevicePage()));
       }
     });
+  }
+
+  Future<void> _checkAndShowChangelog() async {
+    try {
+      final packageInfo = await PackageInfo.fromPlatform();
+      final currentVersion = packageInfo.version;
+      final lastSeenVersion = settings.getLastSeenVersion();
+      
+      if (mounted) {
+        await ChangelogDialog.showIfNeeded(context, currentVersion, lastSeenVersion);
+      }
+      
+      // Update last seen version
+      await settings.setLastSeenVersion(currentVersion);
+    } catch (e) {
+      print('Failed to check changelog: $e');
+    }
   }
 
   @override